requirements.txt
paddleocr==1.5.1
pymupdf==1.24.7
pillow==10.3.0
numpy==1.26.4
requests==2.32.3
